fix init twice error, update atx-agent t0 0.4.1
